UnityTests/Python/requirements.txt
absl-py==0.9.0
astunparse==1.6.3
attrs==19.3.0
cachetools==4.1.1
cattrs==1.0.0
certifi==2020.6.20
chardet==3.0.4
cloudpickle==1.3.0
future==0.18.2
gast==0.3.3
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
gym==0.17.2
# Editable install with no version control (gym-unity==0.23.0)
-e D:/Work/ml-agents-release_12/gym-unity
h5py==2.10.0
idna==2.10
Keras-Preprocessing==1.1.2
Markdown==3.2.2
# Editable install with no version control (mlagents==0.23.0)
-e D:/Work/ml-agents-release_12/ml-agents
# Editable install with no version control (mlagents-envs==0.23.0)
-e D:/Work/ml-agents-release_12/ml-agents-envs
numpy==1.18.0
oauthlib==3.1.0
opt-einsum==3.2.1
Pillow==7.2.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.0
pypiwin32==223
pywin32==228
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
